requirements.txt
streamlit>=1.28.0
openai>=1.3.0
langchain>=0.0.340
faiss-cpu>=1.7.4
pandas>=1.5.0
numpy>=1.24.0
plotly>=5.17.0
Pillow>=10.0.0
sqlite3